// generated/src/app/api-models/al-letter-history.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLetterHistory {
  seqLetterHist: number;
  letterEventId: string;
  letterTemplateId: string;
  seqLetterTempId: number;
  groupTemplateId: string;
  key1: string;
  key2: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}